Use <-> for trigram search with custom threshold
The original implementation of custom thresholds in PgSearch
led to a situation where trigram searches couldn't use
indexes and took too long on large datasets.

Our hope is that using the <-> operator instead of the similarity()
function will permit us to take advantage of pg_trgm gin or gist indexes.

Nora Lin deserves full credit for suggesting that we try switching
to the <-> operator to solve the indexing problem.